requirements.txt
requests~=2.25.1
paho-mqtt~=1.5.1
